ml-service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
transformers>=4.44.0
torch>=2.3.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
shap>=0.45.0
lime>=0.2.0.1
umap-learn>=0.5.6
plotly>=5.22.0
captum>=0.7.0
sentence-transformers>=3.0.0
kaggle>=1.6.14
